Initial attempts at porting the code to custom rule conditions panel
